backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pydantic==2.10.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
python-dotenv==1.0.1
httpx==0.28.1
supabase==2.11.0
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10
